+++ docs/releases.md
@@ -6,7 +6,7 @@ The changes have been tracked starting version 0.1.0
 ## version 0.3.1
 
 * adding the `workspace` module that allow more dynamic and better presentation of the data. [Documentation](https://github.com/pitchmuc/adobe-analytics-api-2.0/blob/master/docs/workspace.md)
-* adding the getReport2 method and [documentation](./main.md##getreport2)
+* adding the getReport2 method and [documentation](./main.md#getreport2)
 * Improving the RequestCreator class by adding 
   * easy load of JSON files
   * capability to setSearch