+++ docs/releases.md
@@ -14,9 +14,9 @@ The changes have been tracked starting version 0.1.0
 * Support RequestCreator instance in the getReport methods
 * adding more scheduleJob methods:
   * getScheduledJob for specific job
-  * createSchedule
-  * createScheduleJob
-  * deleteScheduleJob
+  * createScheduledJob
+  * deleteScheduledJob
+  * updateScheduledJob
   * getDeliverySettings
   * getDeliverySetting
   * createDeliverySetting
@@ -24,6 +24,8 @@ The changes have been tracked starting version 0.1.0
   * deleteDeliverySetting
 * adding [update methods documentation](https://github.com/pitchmuc/adobe-analytics-api-2.0/blob/master/docs/main.md#update-methods)
 * Fixing references of the EventType for Audit Log `LOGS_EVENT_TYPE` attribute.
+Patch: 
+* renaming some scheduleReport method typo.
 
 
 ## version 0.3.0