+++ docs/releases.md
@@ -23,10 +23,10 @@ The changes have been tracked starting version 0.1.0
   * updateDeliverySetting
   * deleteDeliverySetting
 * adding [update methods documentation](https://github.com/pitchmuc/adobe-analytics-api-2.0/blob/master/docs/main.md#update-methods)
-* Fixing references of the EventType for Audit Log `LOGS_EVENT_TYPE` attribute.
+* Fixing references of the EventType for Audit Log `LOGS_EVENT_TYPE` attribute.\
 Patch: 
 * renaming some scheduleReport method typo.
-
+* supporting old report request without statistics info
 
 ## version 0.3.0